state-surgery: validate configs using chainid
Move the instantiation of the providers to earlier
in the script so that they can be used to check the
chain id before doing any slow disk reads of giant json
files.

Create wallets to get access to the `getChainId` method
because it is not available on the provider itself.